a100gpuNvidia/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
streamlit==1.37.1
plotly==5.23.0
requests==2.32.3
python-dotenv==1.0.1
